microservices/gatewayApi/pyproject.toml
[tool.poetry]
name = "gwa-api"
version = "1.0.0"
package-mode = false
description = "Gateway API for multi-tenant configuration on Kong API Gateway"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
werkzeug = ">=2"
ply = "3.10"
cryptography = "38.0.4"
authlib = "0.15.3"
flask-swagger-ui = "3.36.0"
Jinja2 = ">=3"
PyYAML = "6.0.1"
munch = "2.5.0"
boto3 = ">=1.9"
flask = "2.3.3"
flask-compress = "1.4.0"
flask-cors = "3.0.9"
gevent = "^25.5.1"
gunicorn = ">=20.1.0"
python-keycloak = "0.22.0"
requests = "^2.32"
flask-jwt-simple = "0.0.3"
urllib3 = "^2.5.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
pycodestyle = "2.10.0"
pylint = "1.7.2"
python-dotenv = "^0.19.1"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[tool.coverage.run]
relative_files = true
omit = [
    "tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
